Renamed defines.h
